broker: avoid spurious overlay peer warning
Problem: the broker complains of being unable to resolve
its upstream peer when bootstrapped via PMI, but the peer
uri is valid.

Warning: unable to resolve upstream peer [: Name or service not known

The URI is generated by 0mq after concretizing a wildcard bind address.
When ipv6 is enabled, 0mq puts the address in a form that cannot
be parsed by yuarel, nor can it be looked up by getaddrinfo(3), e.g.

tcp://[::ffff:134.9.50.69]:49155

Move the check into boot_config.c so that it is only performed when
the URIs come from a config file.

Fixes #5273
