+++ src/broker/boot_config.c
@@ -13,6 +13,9 @@
 #if HAVE_CONFIG_H
 #include "config.h"
 #endif
+#include <sys/types.h>
+#include <sys/socket.h>
+#include <netdb.h>
 #include <unistd.h>
 #include <stdbool.h>
 #include <assert.h>
@@ -22,6 +25,7 @@
 #include <flux/hostlist.h>
 #include <flux/taskmap.h>
 
+#include "src/common/libyuarel/yuarel.h"
 #include "src/common/libutil/log.h"
 #include "src/common/libutil/errno_safe.h"
 #include "ccan/str/str.h"
@@ -495,6 +499,38 @@ static int set_broker_boot_method_attr (attr_t *attrs, const char *value)
     return 0;
 }
 
+/* Zeromq treats failed hostname resolution as a transient error, and silently
+ * retries in the background, which can make config problems hard to diagnose.
+ * Parse the URI in advance, and if the host portion is invalid, log it.
+ * Ref: flux-framework/flux-core#5009
+ */
+static void warn_of_invalid_host (flux_t *h, const char *uri)
+{
+    char *cpy;
+    struct yuarel u = { 0 };
+    struct addrinfo *result;
+    int e;
+
+    if (!(cpy = strdup (uri))
+        || yuarel_parse (&u, cpy) < 0
+        || !u.scheme
+        || !u.host
+        || !streq (u.scheme, "tcp"))
+        goto done;
+    /* N.B. this URI will be used for zmq_connect(), therefore it must
+     * be a valid peer address, not an interface name or wildcard.
+     */
+    if ((e = getaddrinfo (u.host, NULL, NULL, &result)) == 0) {
+        freeaddrinfo (result);
+        goto done;
+    }
+    log_msg ("Warning: unable to resolve upstream peer %s: %s",
+             u.host,
+             gai_strerror (e));
+done:
+    free (cpy);
+}
+
 int boot_config (flux_t *h, struct overlay *overlay, attr_t *attrs)
 {
     struct boot_conf conf;
@@ -623,6 +659,7 @@ int boot_config (flux_t *h, struct overlay *overlay, attr_t *attrs)
                                       parent_uri,
                                       sizeof (parent_uri)) < 0)
             goto error;
+        warn_of_invalid_host (h, parent_uri);
         if (overlay_set_parent_uri (overlay, parent_uri) < 0) {
             log_err ("overlay_set_parent_uri %s", parent_uri);
             goto error;
